+++ library/src/components/Collection.js
@@ -1,9 +1,26 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
+import axios from 'axios';
+import Book from './ItemBook';
 
 const Collection = () => {
+    const [books, setBooks] = useState([])
+
+    useEffect(() => {
+        const fetchBooks = async () => {
+            const { data } = await axios.get("http://localhost:5000")
+            setBooks(data)
+        }
+        fetchBooks()
+    })
     return (
-        <div>
-            <h1>Books</h1>
+        <div className='collectionContainer'>
+            <h1>Collection</h1>
+            <div className='collectionContent'>
+
+                {books.map((book) => {
+                    return <Book title={book.title} author={book.author} createdAt={book.createdAt} />
+                })}
+            </div>
         </div>
     );
 };
