+++ library/src/styles/components/_collection.scss
@@ -1,0 +1,28 @@
+.collectionContainer {
+    margin-top: 20px;
+
+    h1 {
+        text-decoration: underline;
+        text-align: center;
+    }
+
+    .collectionContent {
+        background-color: $white2;
+        padding: 10px;
+        width: 550px;
+
+        .itemBook {
+            display: flex;
+            justify-content: space-between;
+            align-items: center;
+            display: grid;
+            grid-template-columns: 55% 30% 15%;
+
+            p {
+
+                padding: 3px;
+                border: 1px solid white;
+            }
+        }
+    }
+}
