+++ library/src/styles/index.scss
@@ -1,4 +1,3 @@
 @import "./settings";
 @import "./pages/home";
-@import "./pages/collection"
-// @import "./components/"
+@import "./components/collection";
